--- opendg/loader/__init__.py
@@ -1,2 +1,0 @@
-from opendg.loader.dataloader import DGDataLoader
-from opendg.loader.neighbor_loader import DGNeighborLoader
